+++ packages/web/src/requests/workout.ts
@@ -3,6 +3,6 @@ import { WorkoutModel } from '../models/workoutModels'
 
 export const getWorkout = async(type: string, group: string, equip: string) => {
   const res = await get('/workouts', { type, group, equip })
-  const workout: WorkoutModel[] = JSON.parse(JSON.stringify(res.data))
+  const workout: WorkoutModel[][] = JSON.parse(JSON.stringify(res.data))
   return workout
 }
